+++ packages/http-client/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.0.2] - 2025-08-23
+
+### Changed
+- Update `@fistware/http-core` to v3.0.1 for improved HTTP handling and error management
+
 ## [3.0.1] - 2025-08-15
 
 ### Fixed